+++ src/main/java/com/ironknee/Knee2KneelSpring/dto/stage/StageDTO.java
@@ -1,8 +1,8 @@
 package com.ironknee.Knee2KneelSpring.dto.stage;
 
 import com.ironknee.Knee2KneelSpring.entity.GameEntity;
-import com.ironknee.Knee2KneelSpring.service.stage.StageDepartment;
-import com.ironknee.Knee2KneelSpring.service.stage.StageDifficulty;
+import com.ironknee.Knee2KneelSpring.service.stage_unused.StageDepartment;
+import com.ironknee.Knee2KneelSpring.service.stage_unused.StageDifficulty;
 import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import lombok.AllArgsConstructor;
